requirements.txt
django
djangorestframework
markdown
django-filters
djangorestframework-simplejwt